requirements.txt
fastapi
uvicorn
jinja2
python-multipart
itsdangerous
werkzeug
mysql-connector-python
google-auth
google-auth-oauthlib
requests
python-dotenv
openai
